+++ nmod_poly/mul_classical.c
@@ -30,12 +30,11 @@
 
 // Assumes poly1 and poly2 are not length 0
 void _nmod_poly_mul_classical(mp_ptr res, mp_srcptr poly1, 
-							              ulong len1, mp_srcptr poly2, ulong len2, nmod_t mod)
+							              long len1, mp_srcptr poly2, long len2, nmod_t mod)
 {
    long i;
-      
-   ulong log_len = FLINT_BIT_COUNT(len2);
-   ulong bits = FLINT_BITS - mod.norm;
+   long log_len = FLINT_BIT_COUNT(len2);
+   long bits = FLINT_BITS - (long) mod.norm;
 
    if (2*bits + log_len <= FLINT_BITS)
    {
@@ -72,7 +71,7 @@ void _nmod_poly_mul_classical(mp_ptr res, mp_srcptr poly1,
 void nmod_poly_mul_classical(nmod_poly_t res, 
                          const nmod_poly_t poly1, const nmod_poly_t poly2)
 {
-   ulong len_out;
+   long len_out;
    
    if ((poly1->length == 0) || (poly2->length == 0))  
    {
